+++ packages/qvism-preset/src/recipes/list-header.ts
@@ -12,8 +12,10 @@ const listHeader = defineRecipe({
 
     boxSizing: "border-box",
 
-    paddingInline: vars.base.enabled.root.paddingX,
-    paddingBlock: vars.base.enabled.root.paddingY,
+    paddingLeft: vars.base.enabled.root.paddingX,
+    paddingRight: vars.base.enabled.root.paddingX,
+    paddingTop: vars.base.enabled.root.paddingY,
+    paddingBottom: vars.base.enabled.root.paddingY,
 
     fontSize: vars.base.enabled.root.fontSize,
     lineHeight: vars.base.enabled.root.lineHeight,