+++ packages/qvism-preset/src/recipes/select-box.ts
@@ -25,8 +25,10 @@ export const selectBox = defineSlotRecipe({
       justifyContent: "space-between",
       gap: vars.base.enabled.root.gap,
 
-      paddingInline: vars.base.enabled.root.paddingX,
-      paddingBlock: vars.base.enabled.root.paddingY,
+      paddingLeft: vars.base.enabled.root.paddingX,
+      paddingRight: vars.base.enabled.root.paddingX,
+      paddingTop: vars.base.enabled.root.paddingY,
+      paddingBottom: vars.base.enabled.root.paddingY,
 
       borderRadius: vars.base.enabled.root.cornerRadius,